Added hooks
